Reorganize repository
This is start of the reorganizing into a proposed layout in
http://shortn/_q9qGILflPl. I've kept code change minimal. In subsequent
PRs, we'll gradually implement all of the proposed changes such as
having a shared `data` module and selectable scaling configurations.
So the current state is still a bit inconsistent but this should improve
over time.

Note-worthy changes:

- Support both torch_xla and torchax in one pip installable codebase.
- Scripts in `launcher` can submit both torch_xla and torchax
  workloads to an XPK cluster for large scale training.
- I removed the `stablediffusion2_inference_torchax` example because I'm
  not sure how to fit the inference side of things in a training heavy
  codebase. It looks like the same example is already checked in at
  https://github.com/pytorch/xla/blob/master/experimental/reference_models/sdxl_inference/sdxl.py
- I rolled back the transformer lib dependency until torch_xla model
  unit tests can pass again. We really need CI to prevent bit-rot.

Tested:

```
python3 torchprime/torch_xla_models/tests/test_llama.py

XLA_IR_DEBUG=1 XLA_HLO_DEBUG=1 python3 \
    torchprime/torch_xla_models/train.py \
    torchprime/torch_xla_models/configs/run.json

NUM_SLICES=1 TPU_TYPE=v6e-256 launcher/run_xpk.sh \
   torchprime/torch_xla_models/train.py \
   --dataset_name wikitext \
   --dataset_config_name 'wikitext-2-raw-v1' \
   --output_dir /tmp \
   --cache_dir /tmp \
   --global_batch_size 256 \
   --logging_steps 10 \
   --max_steps 15 \
   --profile_step 5 \
   --model_id 'meta-llama/Meta-Llama-3-8B' \
   --tokenizer_name 'meta-llama/Meta-Llama-3-8B' \
   --block_size 8192 \
   --fsdp full_shard \
   --fsdp_config torchprime/torch_xla_models/configs/fsdp_config.json

python3 torchprime/experimental/torchax_models/run.py --batch_size=16

NUM_SLICES=2 TPU_TYPE=v6e-256 launcher/run_xpk.sh \
    torchprime/experimental/torchax_models/run.py \
    --batch_size=256 \
    --model_type=405B \
    --seqlen=8192 \
    --use_custom_offload=True \
    --use_custom_mesh=True \
    --model_impl=scan \
    --tp=4 \
    --unroll_layers=1
```

Current status of torch_xla model: loss is still NaN

Current status of torchax model: OOMs when creating sharded parameters

    Total hbm usage >= 63.25G:
        reserved        260.12M
        program         unknown size
        arguments        63.00G
