+++ .gitignore
@@ -53,3 +53,6 @@ htmlcov/
 
 # Ignore pytest cache
 .pytest_cache/
+
+# Ignore profiler outputs
+/profile/
