+++ docs/README.md
@@ -1,0 +1,4 @@
+# TorchPrime documentation
+
+This directory should contain model training guides such as how to pick
+the best sharding strategies and how to diagnose out-of-memory (OOM) errors.
