--- llama_training_xla/LICENSE
@@ -1,1 +1,0 @@
-../LICENSE
