--- llama_training_xla/pyproject.toml
@@ -1,29 +1,0 @@
-[build-system]
-requires = ["setuptools>=61.0"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "torchprime"
-version = "0.1.0"
-description = "TorchPrime, reference model implementations for torch-xla."
-readme = "README.md"
-requires-python = ">=3.10"
-license = {file = "LICENSE"}
-authors = [
-    {name = "Jiewen Tan", email = "jwtan@google.com"},
-]
-dependencies = [
-    "accelerate==1.2.1",
-    "transformers==4.47.1",
-    "datasets==3.2.0",
-
-    # `transformers` requires `tensorflow` now.
-    "tensorflow==2.15.0",
-]
-
-[tool.setuptools.packages.find]
-where = ["torchprime"]
-include = ["torchprime", "torchprime.*"]
-
-[tool.setuptools.package-data]
-"torchprime" = ["py.typed"]
