+++ torchprime/data/__init__.py
@@ -1,0 +1,4 @@
+"""
+The data module contains datasets and efficient dataloading implementations.
+"""
+
