+++ torchprime/experimental/torchax_models/README.md
@@ -1,7 +1,10 @@
-# LLama training on torchax + TPU
+# torchax models
 
-This directory demonstrates training of llama model. The model implementation is forked
-at the official reference implementation at https://github.com/meta-llama/llama-models/blob/main/models/llama3/reference_impl/model.py
+## Llama
+
+The `llama` directory demonstrates training of llama model. The model
+implementation is forked at the official reference implementation at
+https://github.com/meta-llama/llama-models/blob/main/models/llama3/reference_impl/model.py
 this file is kept as `model_original.py` for reference.
 
 Then, 3 versions of the model is provided: 
@@ -41,16 +44,32 @@ python run.py --model_impl=<orig|scan|scan_manual>
 
 ## Run on XPK
 
-
-Run `CLUSTER_NAME=foobar PROJECT_ID=something ./run_xpk.sh` will build the dockerfile and launch it on XPK.
-Replace `CLUSTER_NAME` and `PROJECT_ID` with appropiate values.
+Run
+`CLUSTER_NAME=... PROJECT_ID=.. ZONE=... launcher/run_xpk.sh <local command>`
+will build the dockerfile and launch it on XPK.
+Replace `CLUSTER_NAME`, `PROJECT_ID`, and `ZONE` with appropiate values.
 
 
 ## Benchmarks (WIP)
+
 |device| Model size | Batch size | seq length | step time | MFU | NOTEs|
 |-------| ----- | ----- | ----- | ----- | ----- | ---|
 |TPU v6e-8| 8B |        8 |      8192 | 1.7s | 30% | Scan, fsdp, host-offload|
 |TPU v6e-256 x 2| 405B | 256 | 8192 | 46.12s | 25.8% | Scan, fsdp + tp, host-offload|
 
+<!-- TODO: support specifying different XLA flags -->
 
+Llama 3.1 405B on v6e-256 x 2 command:
 
+```sh
+NUM_SLICES=2 TPU_TYPE=v6e-256 launcher/run_xpk.sh \
+    torchprime/experimental/torchax_models/run.py \
+    --batch_size=256 \
+    --model_type=405B \
+    --seqlen=8192 \
+    --use_custom_offload=True \
+    --use_custom_mesh=True \
+    --model_impl=scan \
+    --tp=4 \
+    --unroll_layers=1
+```
