+++ torchprime/experimental/torchax_models/custom_mesh.py
